Switch the empty flag from the keg to the filling.
This is done so the brew can display all kegs where it is currently present.
If the keg itself would hold the information if it is empty or not, the brew could not determine if a full keg holds the
brew or was emptied in between and was filled again with a different brew.