Add the possibillity to generate a print template
An endpoint is added, which returns a pdf file with a print template for the qrcodes and the names of the kegs.
There is no direct link in the frontend included, since there is no user or rights management (yet) and this functions
should only be called by a user who knows what he does, the link for generating the print template has to be called
directly.